Removed yazi from chezmoiignore
Here is the rationale: It's only a few mb, it won't be used if the app
is not installed, and it makes everything sooo much cleaner. I'm talking
in general. Why should I go through the effort of making a template
when all I'm gonna do is check the OS. Much less work to just add it
anyway, my repo is unix only anyway.