Got rid of chezmoi external
Causes too many issues tbh. What if didn't want to sync it? What if I
wanted to use a different link because I didn't have an ssh key? I could
spend hours writing and debugging automation just to solve this problem,
or just do it semi manually with a script.